# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAKBFANTASY",
        "description": "Api SAKBFANTASY Documentation",
        "contact": {
            "email": "info@sakb.com"
        },
        "license": {
            "name": "Sakb",
            "url": "http://sakb-co.com.sa/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/statistics": {
            "post": {
                "tags": [
                    "statistics"
                ],
                "summary": "get all details of statistics players in mathes",
                "operationId": "statistics",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "link_team",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_play",
                        "in": "query",
                        "description": "heigh_price,low_price,heighest_point,lowest_point",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "loc_player",
                        "in": "query",
                        "description": "goalkeeper,defender,line,attacker",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "default": "20"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "default": "0"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/subeldwry": {
            "get": {
                "tags": [
                    "subeldwry_fixtures"
                ],
                "summary": "get all subeldwry",
                "operationId": "subeldwry",
                "parameters": [
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/active_subeldwry": {
            "get": {
                "tags": [
                    "subeldwry_fixtures"
                ],
                "summary": "get current active subeldwry",
                "operationId": "active_subeldwry",
                "parameters": [
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}